contact_form/requirements.txt
Flask
Flask-Mail
Flask-WTF
python-dotenv
email-validator
